# swagger.yaml
...
swagger: "2.0"
basePath: "/api/v0.2"
info:
  title: LEXIS Portal API
  description: The LEXIS Portal API serves as the API proxy to all the interfaces that interact with the parts involved in LEXIS.
  version: "0.2.2021-12-15"
  contact:
    email: diego@cyclops-labs.io
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
...
securityDefinitions:
  APIKeyHeader:
    type: apiKey
    in: header
    name: X-API-KEY
  APIKeyParam:
    type: apiKey
    in: query
    name: api_key
  Keycloak:
    type: oauth2
    flow:  accessCode
    # Need to get clarity on whether these are necessary here or provided with a config file
    authorizationUrl: 'https://138.246.232.245:6443/auth'
    tokenUrl: 'https://138.246.232.245:6443/auth'
security:
  - APIKeyHeader: []
  - APIKeyParam: []
  - Keycloak: []
paths:
...
  /dataset/ssh/sshfsexport:
    post:
      tags:
        - dataSetManagement
        - sshfsexport
      consumes:
        - application/json
      produces:
        - application/json
      summary: Export a directory of the DDI for use via SSHFS
      operationId: CreateSSHFSExport
      description: Export a directory of the DDI for use via SSHFS
      parameters:
        - in: body
          name: parameters
          description: parameters
          required: true
          schema:
            type: object
            required:
              - host
              - pubkey
              - path
            properties:
              host:
                type: string
                example: "141.40.145.102"
              pubkey:
                type: string
              path:
                type: string
      responses:
        '201':
          description: Export created
          schema:
            type: object
            required:
              - user
              - sshfs
            properties:
              user:
                type: string
                example: "qlecvxbndl"
              sshfs:
                type: string
                example: "qlecvxbndl@lexis-test.srv.lrz.de:"
        '400':
          description: This means that there's something wrong in the input parameters and the server couldn't understand the request.
          schema:
            $ref: "#/definitions/ErrorResponse"
        '401':
          description:  This means that the user is not authenticated with keycloak and data transfer can't be triggered unless the user first log in with a valid user
          schema:
            $ref: "#/definitions/ErrorResponse"
        '403':
          description: This means that the resource the user is trying to transfer from or to is not readable or writable by the user. User doesn't have the correct rights to either read the source file or write on the target system location.
          schema:
            $ref: "#/definitions/ErrorResponse"
        '503':
          description: unexpected error
          schema:
            $ref: '#/definitions/ErrorResponse'
    delete:
      tags:
        - dataSetManagement
        - sshfsexport
      consumes:
        - application/json
      produces:
        - application/json
      summary: Export a directory of the DDI for use via SSHFS
      operationId: DeleteSSHFSExport
      description: Remove an SSHFS export of a directory of the DDI
      parameters:
        - in: body
          name: parameters
          description: parameters
          required: true
          schema:
            type: object
            required:
              - user
              - path
            properties:
              user:
                type: string
                example: "qlecvxbndl"
              path:
                type: string
      responses:
        '204':
          description: Export removed
        '400':
          description: This means that there's something wrong in the input parameters and the server couldn't understand the request.
          schema:
            $ref: "#/definitions/ErrorResponse"
        '401':
          description:  This means that the user is not authenticated with keycloak and data transfer can't be triggered unless the user first log in with a valid user
          schema:
            $ref: "#/definitions/ErrorResponse"
        '403':
          description: This means that the resource the user is trying to transfer from or to is not readable or writable by the user. User doesn't have the correct rights to either read the source file or write on the target system location.
          schema:
            $ref: "#/definitions/ErrorResponse"
        '404':
          description: The resource (username, path or both) does not exist
          schema:
           $ref: "#/definitions/ErrorResponse"
        '502':
          description: bad gateway (backend service is down)
          schema:
            $ref: '#/definitions/ErrorResponse'
        '503':
          description: unexpected error
          schema:
            $ref: '#/definitions/ErrorResponse'
...
definitions:
...
  ErrorResponse:
    type: object
    required:
      - errorString
    properties:
      errorString:
        type: string
...